requirements.txt
pandas==1.3.3
yfinance==0.1.63
discord.py==1.7.3
emoji==1.4.2
tabulate==0.8.9
youtube-dl==2021.6.6
requests==2.22.0
PyNaCl==1.3.0
validators==0.18.2